remove old ImageEditor.js file
